pdf_email_app/requirements.txt
PyPDF2
pandas
Pillow
tqdm
pymupdf
customtkinter
